backend/pyproject.toml
[project]
name = "starter-kit"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.15.1",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "crypto>=1.4.1",
    "email-validator>=2.2.0",
    "faker>=37.1.0",
    "fastapi[standard]>=0.115.11",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2>=2.9.10",
    "pydantic-settings>=2.8.1",
    "python-dotenv>=1.0.1",
    "python-jose>=3.4.0",
    "sqlalchemy[asyncio]>=2.0.38",
    "uvicorn>=0.34.0",
]
